Remove redundant and potentially confusing paragraph about the chair
The first sentence of this paragraph repeats some of the text in the
previous paragraph. The second sentence covers powers and
responsibilities shared with the PMC, and stating this in this section
runs the risk of implying that the chair has special powers, which is
not the case. (See § 3.8. Decision Types)